detection-factory/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiv-detection-factory==4.3.7
aivcontracts==1.0.18
aivstorage==1.0.5
async-timeout==4.0.2
attrs==22.2.0
azure-core==1.26.3
azure-storage-blob==12.15.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
fonttools==4.39.2
frozenlist==1.3.3
idna==3.4
importlib-resources==5.12.0
isodate==0.6.1
kafka-python==2.0.2
kiwisolver==1.4.4
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psutil==5.9.4
pycparser==2.21
pyparsing==3.0.9
python-configuration==0.8.2
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
seaborn==0.12.2
scipy==1.9.3
six==1.16.0
thop==0.0.31.post2005241907
torch==1.10.2
torchvision==0.11.3
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
zipp==3.15.0
